requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
python-dotenv==1.0.0
pyodbc==4.0.39
gunicorn==21.2.0
Werkzeug==2.3.7
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
email-validator==2.0.0 